backend/requirements.txt
# Backend Dependencies for Personal AI Assistant

# Core Flask framework
Flask==2.3.3
Flask-CORS==4.0.0

# HTTP requests and API calls
requests==2.31.0
aiohttp==3.8.5

# Environment variable management
python-dotenv==1.0.0

# Google API integrations
google-api-python-client==2.98.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0

# Voice processing
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11

# Data handling and utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and testing
pytest==7.4.2
pytest-flask==1.2.0
python-decouple==3.8

# Security
cryptography==41.0.4

# JSON Web Tokens (if needed for authentication)
PyJWT==2.8.0

# Async support
asyncio==3.4.3

# Logging
colorlog==6.7.0
